+++ apps/etl/extraction/sources/gdacs/extract.py
@@ -1,4 +1,3 @@
-import hashlib
 import json
 import logging
 import typing
@@ -6,9 +5,9 @@ import typing
 import numpy as np
 import pandas as pd
 from celery import shared_task
-from django.core.files.base import ContentFile
 from pydantic import ValidationError
 
+from apps.etl.extraction.sources.base.utils import store_extraction_data
 from apps.etl.extraction.sources.base.extract import Extraction
 from apps.etl.extraction.sources.gdacs.validators.gdacs_eventsdata import (
     GDacsEventDataValidator,
@@ -123,78 +122,6 @@ def validate_gdacs_geometry_data(resp_data):
         "validation_error": validation_error if validation_error else "",
     }
     return validation_data
-
-
-def manage_duplicate_file_content(source, hash_content, instance, response_data, file_name):
-    """
-    if duplicate file content exists then do not create a new file, but point the url to
-    the previous file.
-    """
-    duplicate_file_content = ExtractionData.objects.filter(source=source, file_hash=hash_content)
-    if duplicate_file_content:
-        instance.resp_data = duplicate_file_content.first().resp_data
-        instance.revision_id = duplicate_file_content.first()
-    else:
-        instance.resp_data.save(file_name, ContentFile(response_data))
-    instance.save()
-
-
-def hash_file_content(content):
-    """
-    Compute the hash of a file using the specified algorithm.
-    :return: Hexadecimal hash of the file
-    """
-    file_hash = hashlib.sha256(content).hexdigest()
-    return file_hash
-
-
-def store_extraction_data(
-    response,
-    source=None,
-    validate_source_func=None,
-    parent_id=None,
-    instance_id=None,
-    hazard_type=None,
-    requires_hazard_type=False,
-):
-    file_extension = response.pop("file_extension")
-    file_name = f"gdacs.{file_extension}"
-    resp_data = response.pop("resp_data")
-
-    # save the additional response data after the data is fetched from api.
-    gdacs_instance = ExtractionData.objects.get(id=instance_id)
-    for key, value in response.items():
-        setattr(gdacs_instance, key, value)
-    gdacs_instance.save()
-
-    # save parent id if it is child extraction object
-    if parent_id:
-        gdacs_instance.parent_id = parent_id
-        gdacs_instance.save(update_fields=["parent_id"])
-
-    # Validate the non empty response data.
-    if resp_data and not response["resp_code"] == 204:
-        resp_data_content = resp_data.content
-        # Source validation
-        # if the validate function requires hazard type as argument pass it as argument else don't.
-        if validate_source_func:
-            if requires_hazard_type:
-                gdacs_instance.source_validation_status = validate_source_func(resp_data_content, hazard_type)["status"]
-                gdacs_instance.content_validation = validate_source_func(resp_data_content, hazard_type)["validation_error"]
-            else:
-                gdacs_instance.source_validation_status = validate_source_func(resp_data_content)["status"]
-                gdacs_instance.content_validation = validate_source_func(resp_data_content)["validation_error"]
-
-        # manage duplicate file content.
-        hash_content = hash_file_content(resp_data_content)
-        manage_duplicate_file_content(
-            source=source,
-            hash_content=hash_content,
-            instance=gdacs_instance,
-            response_data=resp_data_content,
-            file_name=file_name,
-        )
-    return gdacs_instance
 
 
 @shared_task(bind=True, max_retries=3, default_retry_delay=5)
@@ -331,7 +258,4 @@ def fetch_gdacs_geometry_data(self, parent_id, footprint_url, **kwargs):
             parent_id=parent_id,
         )
 
-        with gdacs_instance.resp_data.open() as file:
-            data = file.read()
-
-        return {"extraction_id": gdacs_instance.id, "extracted_data": data}
+        return gdacs_instance.id
