+++ apps/etl/management/commands/extract_desinventar_data.py
